==> report/02-Test-Training-Preparation.ipynb <==
# Preparing the Training/Test Set

As we planned to carry out classification with our topic models, we required a test set to be held out. Due to the dataset being very large, we were unsure as to whether we would be able to successfully carry out both training on all of the training set and testing on all of the test set, but decided we could always down sample them at a later stage if we needed to.

import os
import pandas as pd
import random
from numpy.random import RandomState

As the dataset is very large, we cannot load it in automatically by, for example, a google drive link. Therefore, to run any lines which read in a dataset, like the one below, make sure to have the dataset downloaded to your working directory of your machine. This is the whole dataset in its original form from Kaggle.

df = pd.read_csv('Books_rating.csv')

df.head()

We begin by removing the columns which we will not be using in this project. We are primarily focussing on developing topic models and investigating through this how the information gained solely from the reviews impacts a classification task.

df.drop(['Id','Title','Price','User_id','profileName','review/helpfulness','review/time'], axis=1, inplace=True)


df.head()

We can rename the columns as so.

df=df.rename(columns={'review/score': 'score', 'review/summary': 'summary', 'review/text': 'text'})

df_combined=df
df_combined.head()

Now we add the summary to the review text, as we will be comparing how putting both into our models changes the results from just having the summaries.

df_combined['text'] = df.summary.astype(str).str.cat(df.text.astype(str), sep=' ')

df_combined.head()

df_combined.iloc[0,2]

The following section performs stratified sampling to split the data into a test and training set, which correspond to 20% and 80% of the dataset respectively.

rng = RandomState()

ones = df_combined[df_combined['score'] == 1]
twos = df_combined[df_combined['score'] == 2]
threes = df_combined[df_combined['score'] == 3]
fours = df_combined[df_combined['score'] == 4]
fives = df_combined[df_combined['score'] == 5]

num_ones = round(len(ones) * 0.2)
num_twos = round(len(twos) * 0.2)
num_threes = round(len(threes) * 0.2)
num_fours = round(len(fours) * 0.2)
num_fives = round(len(fives) * 0.2)

test_ones = ones.sample(n=num_ones, random_state=rng)
train_ones = ones.loc[~ones.index.isin(test_ones.index)]
test_twos = twos.sample(n=num_twos, random_state=rng)
train_twos = twos.loc[~twos.index.isin(test_twos.index)]
test_threes = threes.sample(n=num_threes, random_state=rng)
train_threes = threes.loc[~threes.index.isin(test_threes.index)]
test_fours = fours.sample(n=num_fours, random_state=rng)
train_fours = fours.loc[~fours.index.isin(test_fours.index)]
test_fives = fives.sample(n=num_fives, random_state=rng)
train_fives = fives.loc[~fives.index.isin(test_fives.index)]

testframe = [test_ones,test_twos,test_threes,test_fours,test_fives]
trainframe = [train_ones,train_twos,train_threes,train_fours,train_fives]

test = pd.concat(testframe)
train = pd.concat(trainframe)

We save the test and train set to csvs with the following names.

test.to_csv('reviews_test.csv', sep=",", index=False)
train.to_csv('reviews_train.csv', sep=",", index=False)

Finally, this is a quick check to show that the test set is in fact 20% of the original dataset.

a = test.shape[0]
b = train.shape[0]
100 * a/(a+b)
