==> report/06-LDA-toy-example.ipynb <==
# Demonstration of LDA Approach

As our dataset is so large, it is not possible to create and test models from it on our laptops. Therefore, we have made use of a high performance computing system to carry out the following processes on the full dataset. To demonstate what has been done, here we downsample the dataset and go through the same steps.

## 1. Downsampling

We begin by loading in our cleaned training datasets, one which contains the combined review text and summary and the other which contains just the summaries. We then perform stratified sampling according to the sentiment of the reviews.

import pandas as pd

import numpy as np
np.random.seed(2018)
import gensim
from gensim import corpora
from numpy.random import RandomState
rng = RandomState()

trainingDf_combined = pd.read_csv('data_cleaned_full_text_and_summaries.csv')

trainingDf_summaries = pd.read_csv('data_cleaned_full_summaries.csv')

trainingDf_combined.head()

trainingDf_summaries.head()

trainingDf_summaries.shape

ones = trainingDf_combined[trainingDf_combined['sentiment'] == 1]
zeros = trainingDf_combined[trainingDf_combined['sentiment'] == 0]

num_ones = round(len(ones) * 1/2.4)
num_zeros = round(len(zeros) * 1/2.4)

sample_ones = ones.sample(n=num_ones, random_state=rng)
sample_zeros = zeros.sample(n=num_zeros, random_state=rng)

sampleframe = [sample_ones,sample_zeros]

train_combined = pd.concat(sampleframe)

train_summaries = trainingDf_summaries.loc[train_combined.index,:]

train_combined.head()

train_summaries.head()

## 2. LDA Model

We now proceed to create an LDA model for the first dataset.

def myprocess(thisdoc):
    ## Runs on documents (vector of words)
    return(thisdoc.strip('[]').replace("'", ' ').replace("u ", '').replace(' ', '').split(','))

docs=train_combined['text'].map(myprocess) 

dict_LoS = corpora.Dictionary(docs)

dict_LoS.filter_extremes(no_below=5, no_above=0.5)

bow_corpus = [dict_LoS.doc2bow(w) for w in docs]

lda_model = gensim.models.LdaMulticore(bow_corpus, num_topics=58, id2word=dict_LoS, passes=1, workers=2)

Now, get the topic distributions for the documents in the training set, and convert them into a dataframe.

lda_scores = lda_model[bow_corpus]

all_topics_csr = gensim.matutils.corpus2csc(lda_scores)
all_topics_numpy = all_topics_csr.T.toarray()
all_topics_pandas = pd.DataFrame(all_topics_numpy)

all_topics_pandas.head()

We want the topic scores to be our X value and our target value, y, is the sentiments of the reviews.

y_train=train_combined['sentiment']
X_train=all_topics_pandas

y_train=pd.DataFrame(y_train, columns=['sentiment'])

## 3. Classification

Now build a Boosted Decision Tree model, making sure to reweight as we have unbalanced classes.

from xgboost import XGBClassifier
import xgboost as xgb

posScaling=(len(y_train['sentiment'])-sum(y_train['sentiment']))/sum(y_train['sentiment'])

model_combined = XGBClassifier(
    objective = 'binary:logistic',
    eval_metric='logloss', 
    scale_pos_weight = posScaling, 
    learning_rate =0.05,
    n_estimators=10000)

Here we use a cross validation method to choose how many trees should be combined into our boosted decision tree.

#xgb_param = model_combined.get_xgb_params()
#xgtrain = xgb.DMatrix(X_train.values, label=y_train.values)

#cvresult = xgb.cv(xgb_param, xgtrain, num_boost_round=model_combined.get_params()['n_estimators'], nfold=3,
                          early_stopping_rounds=10, metrics={'logloss'})


#n_est=cvresult.shape[0]
#print(cvresult.shape[0])
n_est=300

Now, redefine the boosted decision tree model to have that number of trees, and fit it to the training data.

model_combined = XGBClassifier(
    objective = 'binary:logistic',
    eval_metric='logloss', 
    scale_pos_weight = posScaling, 
    learning_rate =0.05,
    n_estimators=n_est)

model_combined.fit(X_train, y_train)

We complete a similar process for the dataset with just summaries. We will just choose to define the boosted decision tree model the same as above in order to save time.

model_summaries = XGBClassifier(
    objective = 'binary:logistic',
    eval_metric='logloss', 
    scale_pos_weight = posScaling, 
    learning_rate =0.05,
    n_estimators=n_est)

docs_summaries=train_summaries['text'].map(myprocess) 
dict_LoS_summaries = corpora.Dictionary(docs_summaries)
dict_LoS_summaries.filter_extremes(no_below=5, no_above=0.5)
bow_corpus_summaries = [dict_LoS_summaries.doc2bow(w) for w in docs_summaries]
lda_model_summaries = gensim.models.LdaMulticore(bow_corpus_summaries, num_topics=58, id2word=dict_LoS_summaries, passes=1, workers=2)
lda_scores_summaries = lda_model_summaries[bow_corpus_summaries]

all_topics_csr_summaries = gensim.matutils.corpus2csc(lda_scores_summaries)
all_topics_numpy_summaries = all_topics_csr_summaries.T.toarray()
all_topics_pandas_summaries = pd.DataFrame(all_topics_numpy_summaries)

X_train_summaries=all_topics_pandas_summaries

X_train_summaries

y_train

model_summaries.fit(X_train_summaries, y_train)

Now load in the test sets. We begin by processing them in a similar way to the training sets, turning them into a bag of words, but using the corresponding dictionaries produced previously. Then, compute topic scores for each of the documents using the LDA models created above, and finally turn these into a dataframe as we did before.

testDf_combined = pd.read_csv('test_cleaned_text_and_summaries.csv')
testDf_summaries = pd.read_csv('test_cleaned_summaries.csv')

docs_combined_test=testDf_combined['text'].map(myprocess) 
bow_corpus_combined_test = [dict_LoS.doc2bow(w) for w in docs_combined_test]
lda_scores_combined_test = lda_model[bow_corpus_combined_test]
all_topics_csr_combined_test = gensim.matutils.corpus2csc(lda_scores_combined_test)
all_topics_numpy_combined_test = all_topics_csr_combined_test.T.toarray()
all_topics_pandas_combined_test = pd.DataFrame(all_topics_numpy_combined_test)
X_test_combined=all_topics_pandas_combined_test

docs_summaries_test=testDf_summaries['text'].map(myprocess) 
bow_corpus_summaries_test = [dict_LoS_summaries.doc2bow(w) for w in docs_summaries_test]
lda_scores_summaries_test = lda_model_summaries[bow_corpus_summaries_test]
all_topics_csr_summaries_test = gensim.matutils.corpus2csc(lda_scores_summaries_test)
all_topics_numpy_summaries_test = all_topics_csr_summaries_test.T.toarray()
all_topics_pandas_summaries_test = pd.DataFrame(all_topics_numpy_summaries_test)
X_test_summaries=all_topics_pandas_summaries_test

## 4. Prediction for test set

We are now ready to compute our prediction probabilities for the test set.

pred_probs_combined = model_combined.predict_proba(X_test_combined)
pred_probs_summaries = model_summaries.predict_proba(X_test_summaries)

pred_probs_combined_df=pd.DataFrame(pred_probs_combined[:,1],columns=['Probability of positive'])
pred_probs_combined_df.to_csv('pred_probs_combined_1M.csv', index=False)

pred_probs_summaries_df=pd.DataFrame(pred_probs_summaries[:,1],columns=['Probability of positive'])
pred_probs_summaries_df.to_csv('pred_probs_summaries_1M.csv', index=False)

import pickle
pickle.dump(model_combined, open(model_combined.pkl, 'wb'))
pickle.dump(model_summaries, open(model_summaries.pkl, 'wb'))

train_summaries.to_csv('train_summaries_1M.csv', index=False)
train_combined.to_csv('train_combined_1M.csv', index=False)
